requirements.txt
# FastAPI Backend
fastapi>=0.100
uvicorn>=0.20
pydantic>=2.0

# ForkLens Pipeline
torch>=2.0
transformers>=4.0
sentence-transformers>=2.0
qdrant-client>=1.6
openai>=1.0
python-dotenv>=1.0

# Fallback (optional)
streamlit>=1.20